add separate utils file for non-react stuff
