+++ glitch-dot-cool/src/pages/tags.js
@@ -2,7 +2,8 @@ import React from "react"
 import { graphql, useStaticQuery } from "gatsby"
 
 import Layout from "../components/layout"
-import { ListLink, StyledLink } from "../utils/utils"
+import { ListLink, StyledLink } from "../utils/utilComponents"
+import { slugify } from "../utils/utils"
 
 const Tags = () => {
   const data = useStaticQuery(graphql`
@@ -36,20 +37,3 @@ const Tags = () => {
 }
 
 export default Tags
-
-const slugify = string => {
-  const a = "àáäâãåăæçèéëêǵḧìíïîḿńǹñòóöôœøṕŕßśșțùúüûǘẃẍÿź·/_,:;"
-  const b = "aaaaaaaaceeeeghiiiimnnnooooooprssstuuuuuwxyz------"
-  const p = new RegExp(a.split("").join("|"), "g")
-
-  return string
-    .toString()
-    .toLowerCase()
-    .replace(/\s+/g, "-") // Replace spaces with -
-    .replace(p, c => b.charAt(a.indexOf(c))) // Replace special characters
-    .replace(/&/g, "-and-") // Replace & with 'and'
-    .replace(/[^\w\-]+/g, "") // Remove all non-word characters
-    .replace(/\-\-+/g, "-") // Replace multiple - with single -
-    .replace(/^-+/, "") // Trim - from start of text
-    .replace(/-+$/, "") // Trim - from end of text
-}
